refactor: adapters to their own folder
