+++ src/adapters/implementations/BearyChatResponder.js
@@ -1,5 +1,5 @@
 'strict';
-const { Responder } = require('./Responder');
+const { Responder } = require('../Responder');
 
 class BearyChatResponder extends Responder {
   /**
